requirements.txt
aiohttp==3.9.5
beautifulsoup4==4.12.3
Pillow==10.1.0
Pillow==10.4.0
Requests==2.32.3
rich==13.8.1
